Moving the worker responsible for sending notifications into the web
