--- tracker/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["es2015", "node6", "react"]
-}
-
