--- tracker/package.json
@@ -1,81 +1,0 @@
-{
-  "name": "web",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@slack/client": "^3.10.0",
-    "autoprefixer": "7.1.0",
-    "babel-cli": "^6.24.1",
-    "babel-core": "6.24.1",
-    "babel-eslint": "7.2.3",
-    "babel-jest": "20.0.3",
-    "babel-loader": "7.0.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-node6": "^11.0.0",
-    "babel-preset-react-app": "^3.0.0",
-    "babel-runtime": "6.23.0",
-    "body-parser": "^1.17.2",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
-    "cfenv": "^1.0.4",
-    "chalk": "1.1.3",
-    "cloudant": "^1.8.0",
-    "css-loader": "0.28.1",
-    "eslint": "3.19.0",
-    "eslint-config-react-app": "^1.0.4",
-    "eslint-loader": "1.7.1",
-    "eslint-plugin-flowtype": "2.33.0",
-    "eslint-plugin-import": "2.2.0",
-    "eslint-plugin-jsx-a11y": "5.0.3",
-    "eslint-plugin-react": "7.0.1",
-    "express": "^4.15.3",
-    "extract-text-webpack-plugin": "2.1.0",
-    "file-loader": "0.11.1",
-    "fs-extra": "3.0.1",
-    "gifencoder": "^1.0.6",
-    "html-webpack-plugin": "2.28.0",
-    "jest": "20.0.3",
-    "morgan": "^1.8.2",
-    "nodemailer": "^4.0.1",
-    "object-assign": "4.1.1",
-    "phantom": "^4.0.3",
-    "png-file-stream": "^1.0.0",
-    "postcss-flexbugs-fixes": "3.0.0",
-    "postcss-loader": "2.0.5",
-    "promise": "7.1.1",
-    "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "react-bootstrap": "^0.31.0",
-    "react-dev-utils": "^3.0.0",
-    "react-dom": "^15.5.4",
-    "react-error-overlay": "^1.0.7",
-    "react-router-dom": "^4.1.1",
-    "recharts": "^1.0.0-alpha.0",
-    "request": "^2.81.0",
-    "style-loader": "0.17.0",
-    "sw-precache-webpack-plugin": "0.9.1",
-    "tiny-worker": "^2.1.1",
-    "url-loader": "0.5.8",
-    "url-search-params-polyfill": "^1.3.0",
-    "watson-developer-cloud": "^2.32.1",
-    "watson-react-components": "^0.6.9",
-    "webpack": "2.6.1",
-    "webpack-dev-server": "2.4.5",
-    "webpack-manifest-plugin": "1.1.0",
-    "whatwg-fetch": "^2.0.3"
-  },
-  "engines": {
-    "node": "7.10.0"
-  },
-  "scripts": {
-    "start": "./node_modules/babel-cli/bin/babel-node.js ./worker.js",
-    "build": "./node_modules/babel-cli/bin/babel-node.js ./build.js"
-  },
-  "babel": {
-    "presets": [
-      "react-app"
-    ]
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  }
-}
