+++ web/package.json
@@ -34,6 +34,7 @@
     "html-webpack-plugin": "2.28.0",
     "jest": "20.0.3",
     "morgan": "^1.8.2",
+    "nodemailer": "^4.0.1",
     "object-assign": "4.1.1",
     "postcss-flexbugs-fixes": "3.0.0",
     "postcss-loader": "2.0.5",
@@ -64,6 +65,8 @@
   },
   "scripts": {
     "startdev": "./node_modules/babel-cli/bin/babel-node.js ./scripts/start.js",
+    "start-notifier": "yarn run build-notifier && ./node_modules/babel-cli/bin/babel-node.js ./notifier.js",
+    "build-notifier": "./node_modules/babel-cli/bin/babel-node.js ./src/buildNotifierDB.js",
     "start": "./node_modules/babel-cli/bin/babel-node.js ./server.js",
     "build": "./node_modules/babel-cli/bin/babel-node.js ./scripts/build.js",
     "test": "./node_modules/babel-cli/bin/babel-node.js ./scripts/test.js --env=jsdom"
